added pre-commit and model files, refactored files and archi
